// frontend/src/api/generated/chatbot-client/api.ts
export const ChatConversationsResponseFinishReasonEnum = {
  FinishReasonUnspecified: "FINISH_REASON_UNSPECIFIED",
  FinishReasonStop: "FINISH_REASON_STOP",
  FinishReasonLength: "FINISH_REASON_LENGTH",
  FinishReasonSensitive: "FINISH_REASON_SENSITIVE",
  FinishReasonContext: "FINISH_REASON_CONTEXT",
  FinishReasonUnknown: "FINISH_REASON_UNKNOWN",
} as const;

export type ChatConversationsResponseFinishReasonEnum =
  (typeof ChatConversationsResponseFinishReasonEnum)[keyof typeof ChatConversationsResponseFinishReasonEnum];
// ...
export const FeedBackChatHistoryRequestTypeEnum = {
  FeedBackTypeUnspecified: "FEED_BACK_TYPE_UNSPECIFIED",
  FeedBackTypeLike: "FEED_BACK_TYPE_LIKE",
  FeedBackTypeUnlike: "FEED_BACK_TYPE_UNLIKE",
} as const;

export type FeedBackChatHistoryRequestTypeEnum =
  (typeof FeedBackChatHistoryRequestTypeEnum)[keyof typeof FeedBackChatHistoryRequestTypeEnum];